+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - fixed preserving order of inherited types
 - improved support for throwing methods in protocols
+- fixed extracting parameters of methods with closures in their bodies
 
 ### Internal changes